Rename rovers to dodgy.
rover is already a crate, pluralising just seems malicious. There is a
dodge crate, but it has been empty for almost a year now and I think
dodgy is different enough anyway.

Also dodgy for a crate sounds funny to me. Maybe it'll bite me in the
ass in the future, who knows.