Resolve panic for invalidating obstacles by falling back to a zero velocity.
In the original paper, they choose a velocity of zero for the obstacle
constraints so that the agent can trivially solve the obstacle
constraints. However, we use the agent's velocity for this because
otherwise an agent does weird "wiggling" around obstacle vertices due to
the constraint rapidly changing direction. This is great visually, but
can occasionally lead to panics. So we solve this by falling back to the
zero velocity if the obstacle constraints are violated.